Move config cli logic to a separate file
